# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Application API overview",
    "version": "v1.beta.1.0.0"
  },
  "paths": {
    "/api/user/login": {
      "post": {
        "description": "login user .  Duplicates are not allowed",
        "operationId": "login",
        "tags": ["user"],
        "parameters": [
          {
            "name": "login",
            "in": "body",
            "description": "login user and get token",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "login response",
            "schema": {
              "$ref": "#/definitions/LoginRes"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/api/user/register": {
      "post": {
        "description": "register user .  Duplicates are not allowed",
        "operationId": "register",
        "tags": ["user"],
        "parameters": [
          {
            "name": "register",
            "in": "body",
            "description": "register user and get token",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Register"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "register response",
            "schema": {
              "$ref": "#/definitions/RegisterRes"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/api/books": {
      "get": {
        "operationId": "listVersionsv2",
        "summary": "Get All Books",
        "tags": ["books"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "200 response",
            "examples": {
              "application/json": {
                "items": [
                  {
                    "_id": "62798d25c8b987241f7a4a25",
                    "name": "name updated!"
                  },
                  {
                    "_id": "627acd075ff8ea210182a622",
                    "name": "send new book from post man new Edited!"
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "description": "Creates a new book .  Duplicates are allowed",
        "operationId": "addBook",
        "tags": ["books"],
        "parameters": [
          {
            "name": "book",
            "in": "body",
            "description": "Book to add to the Database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewBook"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "book response",
            "schema": {
              "$ref": "#/definitions/Book"
            }
          },
          "default": {
            "description": "unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Login": {
      "type": "object",
      "allOf": [
        {
          "$ref": "#/definitions/Login"
        },
        {
          "required": ["email", "password"],
          "properties": {
            "email": {
              "type": "string",
              "format": "int64"
            },
            "password": {
              "type": "string",
              "format": "int64"
            }
          }
        }
      ]
    },
    "LoginRes": {
      "type": "object",
      "allOf": [
        {
          "$ref": "#/definitions/LoginRes"
        },
        {
          "required": ["email", "username", "token"],
          "properties": {
            "email": {
              "type": "string",
              "format": "int64"
            },
            "username": {
              "type": "string",
              "format": "int64"
            },
            "token": {
              "type": "string",
              "format": "int64"
            }
          }
        }
      ]
    },
    "Register": {
      "type": "object",
      "allOf": [
        {
          "$ref": "#/definitions/Register"
        },
        {
          "required": ["email", "password"],
          "properties": {
            "email": {
              "type": "string",
              "format": "int64"
            },
            "password": {
              "type": "string",
              "format": "int64"
            }
          }
        }
      ]
    },
    "RegisterRes": {
      "type": "object",
      "allOf": [
        {
          "$ref": "#/definitions/RegisterRes"
        },
        {
          "required": ["email", "username", "token"],
          "properties": {
            "email": {
              "type": "string",
              "format": "int64"
            },
            "username": {
              "type": "string",
              "format": "int64"
            },
            "token": {
              "type": "string",
              "format": "int64"
            }
          }
        }
      ]
    },
    "Book": {
      "type": "object",
      "allOf": [
        {
          "$ref": "#/definitions/NewBook"
        },
        {
          "required": ["id"],
          "properties": {
            "id": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      ]
    },
    "NewBook": {
      "type": "object",
      "required": ["name"],
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "Error": {
      "type": "object",
      "required": ["code", "message"],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "consumes": ["application/json"]
}